# mental-health-frontend-min (2)/微信接口.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "微信接口",
    "description": "前端 API 接口文档转换为 OpenAPI 3.0 格式，用于导入 Apifox。",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/self-help/like": {
      "post": {
        "summary": "点赞文章/内容",
        "deprecated": false,
        "description": "对文章进行点赞\n代码位置: `pages/self-help/article.js` - `handleLike()`",
        "operationId": "likeArticle",
        "tags": [
          "微信前端/内容互动",
          "内容互动"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "currentStatus": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/self-help/collect": {
      "post": {
        "summary": "收藏文章/内容",
        "deprecated": false,
        "description": "对文章进行收藏\n代码位置: `pages/self-help/article.js` - `handleCollect()`",
        "operationId": "collectArticle",
        "tags": [
          "微信前端/内容互动",
          "内容互动"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "currentStatus": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/self-help/list": {
      "get": {
        "summary": "获取内容列表",
        "deprecated": false,
        "description": "获取心理自助内容列表（文章/视频/课程）\n代码位置: `pages/self-help/self-help.js` - `loadContentList()`",
        "operationId": "getSelfHelpList",
        "tags": [
          "微信前端/自助内容展示",
          "自助内容展示"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string",
              "example": "article|video|music|course"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/self-help/detail": {
      "get": {
        "summary": "获取文章详情",
        "deprecated": false,
        "description": "获取文章/视频/课程详情\n代码位置: `pages/self-help/article.js` - `loadArticle()`",
        "operationId": "getSelfHelpDetail",
        "tags": [
          "微信前端/自助内容展示",
          "自助内容展示"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}